Added concurrency capabilities to scriptotron; it's now possible for a scriptotron command callable to include 'yield' statements, which will cause the command to pause execution until Enso's next timer loop.
Interestingly, the addition of this code uncovered a bug in a strange
"fix" to a memory leak bug I made several weeks ago to the OSX
backend, when the Enso codebase was in a different HG repository:

  http://hg.toolness.com/enso_osx/rev/9f5a0cd8a3f0

See the TODO in this commit for more information on this bug.  I still
don't understand the root cause of this bug, but the change I made to
the OSX backend in this commit seems to have fixed it for now. :/